Polish fix for the eb table flow v1.0

+++ sales_report_detailed.asp
@@ -282,22 +282,27 @@
 
                     salesOrderReportFile = "sales_report_container.dbf"
                     prevSalesReportPath = mainPath & "tbl_blank\" & salesOrderReportFile
-                    salesReportPath = mainPath & "temp_folder\" & salesOrderReportFile
+                    salesReportPath = tempFolderPath & salesOrderReportFile
 
-                    if fs.FileExists(salesReportPath) then
+                    if fs.FolderExists(tempFolderPath) <> true then
+                        fs.CreateFolder(tempFolderPath)
+                    end if
 
-                        On Error Resume Next
+                    On Error Resume Next
+
+                        if fs.FileExists(salesReportPath) then
+                        
                             fs.DeleteFile(salesReportPath)
-                        On Error GoTo 0
-                    end if
+                            
+                        end if
 
+                        if fs.FileExists(salesReportPath) <> true then 
 
-                    if fs.FileExists(salesReportPath) <> true then 
+                            fs.CopyFile prevSalesReportPath, salesReportPath
+                        
+                        end if
 
-                        fs.CopyFile prevSalesReportPath, salesReportPath
-                        ' Response.Write "File successfully copied"
-                    
-                    end if
+                    On Error GoTo 0
 
                     for i=0 to monthsDiff
 
@@ -340,11 +345,14 @@
                         if insertSalesReport <> "" then
                             cnroot.execute(insertSalesReport)
                         end if
-                        'Displaying reports'
-                        rs.Open "SELECT cust_id, cust_name, invoice_no, date, prod_gen, prod_price, prod_qty, prodamount FROM "&salesReportPath&" WHERE session_id="&userSessionID&" and duplicate!='yes' and date between CTOD('"&queryDate1&"') and CTOD('"&queryDate2&"') ORDER BY cust_name, cust_id, invoice_no", CN2
                     %>
+                        <%if fs.FileExists(salesReportPath) = true then
 
-                            <%do until rs.EOF
+                            'Displaying reports'
+                            rs.Open "SELECT cust_id, cust_name, invoice_no, date, prod_gen, prod_price, prod_qty, prodamount FROM "&salesReportPath&" WHERE session_id="&userSessionID&" and duplicate!='yes' and date between CTOD('"&queryDate1&"') and CTOD('"&queryDate2&"') ORDER BY cust_name, cust_id, invoice_no", CN2
+                    
+
+                            do until rs.EOF
                                 
                                 Response.Flush
                                 myDate = CDATE(rs("date"))
@@ -431,10 +439,9 @@
                                 <%rs.MoveNext%>
                       
                             <%loop%>
-                        <%rs.close%>   
-
-                        <%
-                            if invoiceCounter > 1 then
+                            <%rs.close%>   
+                        
+                            <%if invoiceCounter > 1 then
                                 if isTotalPrinted = false then%>
                                     <tr> 
                                         <td></td>   
@@ -447,34 +454,23 @@
                                     </tr>
                                 <%end if%>
                             <%end if%>
+
+                        <%end if%>
                           
                 </table>
 
                 <!-- DELETING sales_report_container -->
                 <%  
-                    deleteReport = "DELETE FROM "&salesReportPath&" WHERE session_id="&userSessionID
-                    set objAccess = cnroot.execute(deleteReport)
-                    set objAccess = nothing
+                    if fs.FileExists(salesReportPath) = true then
+
+                        deleteReport = "DELETE FROM "&salesReportPath&" WHERE session_id="&userSessionID
+                        set objAccess = cnroot.execute(deleteReport)
+                        set objAccess = nothing
+
+                    end if
 
                     set rs = nothing
                     CN2.close
-
-
-                    ' closeTbl = "USE IN "&salesReportPath
-                    ' cnroot.execute(closeTbl)
-
-                    
- 
-                     '''Check if file exists before deleting
-                    ' if fs.FileExists(salesReportPath) then
-
-                    '     fs.DeleteFile(salesReportPath)
-                    '     Response.Write "File was deleted"
-
-                    ' end if
-                    ' set fs=nothing
-
-                    
                 %>
 
             </div>
