+++ sales_report_sum_day.asp
@@ -262,22 +262,27 @@
 
                     salesOrderReportFile = "sales_report_container.dbf"
                     prevSalesReportPath = mainPath & "tbl_blank\" & salesOrderReportFile
-                    salesReportPath = mainPath & "temp_folder\" & salesOrderReportFile
+                    salesReportPath = tempFolderPath & salesOrderReportFile
 
-                    if fs.FileExists(salesReportPath) then
+                    if fs.FolderExists(tempFolderPath) <> true then
+                        fs.CreateFolder(tempFolderPath)
+                    end if
 
-                        On Error Resume Next
+                    On Error Resume Next
+
+                        if fs.FileExists(salesReportPath) then
+
                             fs.DeleteFile(salesReportPath)
-                        On Error GoTo 0
-                    end if
+                            
+                        end if
 
+                        if fs.FileExists(salesReportPath) <> true then 
 
-                    if fs.FileExists(salesReportPath) <> true then 
+                            fs.CopyFile prevSalesReportPath, salesReportPath
 
-                        fs.CopyFile prevSalesReportPath, salesReportPath
-                        ' Response.Write "File successfully copied"
-                    
-                    end if
+                        end if
+
+                    On Error GoTo 0
 
                     for i=0 to monthsDiff
 
@@ -320,24 +325,27 @@
                         if insertSalesReport <> "" then
                             cnroot.execute(insertSalesReport)
                         end if
-                        'Displaying reports'
-                        rs.Open "SELECT date, SUM(prodamount) AS prodamount, payment FROM "&salesReportPath&" WHERE session_id="&userSessionID&" and duplicate!='yes' and date between CTOD('"&queryDate1&"') and CTOD('"&queryDate2&"') GROUP BY date, payment ORDER BY date", CN2
-
-                        Dim totalSales, totalCredit, totalCash
-                        totalSales = 0
-                        totalCredit = 0
-                        totalCash = 0
-
-                        Dim dayTotalSales, dayTotalCash, dayTotalCredit
-                        dayTotalSales = 0
-                        dayTotalCash = 0
-                        dayTotalCredit = 0
-
-                        Dim isTotalPrinted, dateCounter
-                        dateCounter = 1
                     %>
+                        <%if fs.FileExists(salesReportPath) = true then
 
-                            <%do until rs.EOF
+                            'Displaying reports'
+                            rs.Open "SELECT date, SUM(prodamount) AS prodamount, payment FROM "&salesReportPath&" WHERE session_id="&userSessionID&" and duplicate!='yes' and date between CTOD('"&queryDate1&"') and CTOD('"&queryDate2&"') GROUP BY date, payment ORDER BY date", CN2
+
+                            Dim totalSales, totalCredit, totalCash
+                            totalSales = 0
+                            totalCredit = 0
+                            totalCash = 0
+
+                            Dim dayTotalSales, dayTotalCash, dayTotalCredit
+                            dayTotalSales = 0
+                            dayTotalCash = 0
+                            dayTotalCredit = 0
+
+                            Dim isTotalPrinted, dateCounter
+                            dateCounter = 1
+                    
+
+                            do until rs.EOF
                                 
                                 Response.Flush
 
@@ -400,27 +408,32 @@
 
                             <%loop%>
 
-                        <%rs.close%>   
+                            <%rs.close%>   
 
-                        <%
-                        if dateCounter > 1 then  
-                            if isTotalPrinted = false then%>
-                                <tr class="final-total">   
-                                    <td>Total</td>   
-                                    <td class="totalAmount">&#8369; <%=totalSales%></td>      
-                                    <td class="totalAmount">&#8369; <%=totalCash%></td>   
-                                    <td class="totalAmount">&#8369; <%=totalCredit%></td>
-                                </tr>
+                            <%if dateCounter > 1 then  
+                                if isTotalPrinted = false then%>
+                                    <tr class="final-total">   
+                                        <td>Total</td>   
+                                        <td class="totalAmount">&#8369; <%=totalSales%></td>      
+                                        <td class="totalAmount">&#8369; <%=totalCash%></td>   
+                                        <td class="totalAmount">&#8369; <%=totalCredit%></td>
+                                    </tr>
+                                <%end if%>
                             <%end if%>
+
                         <%end if%>
                           
                 </table>
 
                 <!-- DELETING sales_report_container -->
                 <%  
-                    deleteReport = "DELETE FROM "&salesReportPath&" WHERE session_id="&userSessionID
-                    set objAccess = cnroot.execute(deleteReport)
-                    set objAccess = nothing
+                    if fs.FileExists(salesReportPath) = true then
+
+                        deleteReport = "DELETE FROM "&salesReportPath&" WHERE session_id="&userSessionID
+                        set objAccess = cnroot.execute(deleteReport)
+                        set objAccess = nothing
+
+                    end if
 
                     set rs = nothing
                     CN2.close
